requirements/base.txt
Django>=5.0,<5.1
djangorestframework>=3.14.0
django-environ>=0.11.0
psycopg2-binary>=2.9.9
redis>=5.0.0
celery>=5.3.0
djangorestframework-simplejwt>=5.3.0
drf-spectacular>=0.27.0
django-cors-headers>=4.3.0
Pillow>=10.1.0
kavenegar>=1.1.0
python-jose>=3.3.0
django-redis>=5.4.0
